feat : article 예외 발생 기능 구현

+++ src/main/java/org/example/article/domain/service/ArticleService.java
@@ -2,6 +2,7 @@ package org.example.article.domain.service;
 
 import org.example.article.domain.entity.Article;
 import org.example.article.domain.entity.Articles;
+import org.example.article.domain.exeption.EntityNotFoundException;
 
 public class ArticleService {
 
@@ -12,12 +13,16 @@ public class ArticleService {
     }
 
     public Article loadArticle(Long id) {
-        return articleRepository.findById(id);
+        return findByIdOrElseThrow(id);
     }
 
     public void update(Long id, String title, String content) {
-        Article article = articleRepository.findById(id);
+        Article article = findByIdOrElseThrow(id);
         article.update(title, content);
+    }
+
+    private Article findByIdOrElseThrow(Long id) {
+        return articleRepository.findById(id).orElseThrow(() -> new EntityNotFoundException("Article not found"));
     }
 
     public Articles loadAll() {
@@ -25,7 +30,7 @@ public class ArticleService {
     }
 
     public void delete(Long id) {
+        if (!articleRepository.extractById(id)) throw new EntityNotFoundException("Article not found");
         articleRepository.deleteById(id);
-        return ;
     }
 }
